@@ -1,11 +1,7 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "./components/app/App";
 import "./style/style.scss";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById("root")
-);
+// StrictMode was temporarily removed (because of React 18+)
+ReactDOM.createRoot(document.getElementById("root")).render(<App />);
