@@ -20,7 +20,7 @@ const App = () => {
     <div className="app">
       <AppHeader />
       <main>
-        {/* <ErrorBoundary>
+        <ErrorBoundary>
           <RandomChar />
         </ErrorBoundary>
         <div className="char__content">
@@ -31,11 +31,11 @@ const App = () => {
             <CharInfo charId={selectedChar} />
           </ErrorBoundary>
         </div>
-        <img className="bg-decoration" src={decoration} alt="vision" /> */}
-        <AppBanner />
+        <img className="bg-decoration" src={decoration} alt="vision" />
+        {/* <AppBanner />
         <ErrorBoundary>
           <ComicsList />
-        </ErrorBoundary>
+        </ErrorBoundary> */}
       </main>
     </div>
   );
